Rename gem to jrmonitoring
Fix version numbering issues with version.rb not being detected
