+++ jrmonitor.gemspec
@@ -1,20 +1,20 @@
 # coding: utf-8
-lib = File.expand_path('../lib', __FILE__)
+lib = File.expand_path('./lib', __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
 require 'monitoring/version'
 
 Gem::Specification.new do |spec|
-  spec.name          = "jruby-monitoring"
-  spec.version       = JRMonitor::VERSION
-  spec.authors       = ["Pere Urbon-Bayes", "Elastic"]
-  spec.email         = ["pere.urbon@gmail.com", "info@elastic.co"]
+  spec.name          = "jrmonitor"
+  spec.version       = "0.4.1"
+  spec.authors       = ["Elastic"]
+  spec.email         = ["info@elastic.co"]
 
   spec.summary       = %q{JVM Platform MXBeans wrapper used for monitoring.}
   spec.description   = %q{This gems allows you to access in a Ruby friendly way the internal JVM monitoring tools for things like System, Threads and Memory information}
   spec.homepage      = "http://www.elastic.co/guide/en/logstash/current/index.html"
   spec.license       = "Apache License (2.0)"
 
-  spec.files         = Dir.glob(["jruby-monitoring.gemspec", "lib/**/*.rb", "spec/**/*.rb", "jruby-ext/target/classes/**/*.class"])
+  spec.files         = Dir.glob(["jrmonitor.gemspec", "lib/**/*.rb", "spec/**/*.rb", "jruby-ext/target/classes/**/*.class"])
   spec.bindir        = "exe"
   spec.executables   = spec.files.grep(%r{^exe/}) { |f| File.basename(f) }
   spec.require_paths = ["lib", "jruby-ext"]
