--- lib/monitoring/version.rb
@@ -1,3 +1,0 @@
-module JRMonitor
-  VERSION = "0.3.1"
-end
